Removed Attack Rate, removed Player class and moved Stats to Hero folder. Created Number Generator.
