+++ DungeonGame/Helpers/NumberGenerator.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DungeonGame.Helpers
+{
+    class NumberGenerator
+    {
+        private readonly Random _random = new Random();
+
+        public int RandomNumber(int min, int max)
+        {
+            return _random.Next(min, max);
+        }
+    }
+}
